requirements.txt
#Django dependencies
Django==2.1.11
django-widget-tweaks==1.4.3
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
#Other dependencies
requests==2.21.0
sentry-sdk==0.10.2
